discord-bot-gemini/requirements.txt
# Core dependencies for Discord LLM Chatbot
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.28.0
aiofiles>=0.23.0
google-generativeai